+++ src/GitHub_Updater/API/Language_Pack_API.php
@@ -85,7 +85,7 @@ class Language_Pack_API extends API {
 				$response = $this->api( '/2.0/repositories/' . $headers['owner'] . '/' . $headers['repo'] . '/src/master/language-pack.json' );
 				break;
 			case 'gitlab':
-				$id       = urlencode( $headers['owner'] . '/' . $headers['repo'] );
+				$id       = rawurlencode( $headers['owner'] . '/' . $headers['repo'] );
 				$response = $this->api( '/projects/' . $id . '/repository/files/language-pack.json' );
 				$response = isset( $response->content )
 					? json_decode( base64_decode( $response->content ) )