+++ src/GitHub_Updater/Plugin.php
@@ -95,10 +95,6 @@ class Plugin extends Base {
 			foreach ( (array) static::$extra_headers as $value ) {
 				$header = null;
 
-				if ( in_array( $value, [ 'Requires PHP', 'Requires WP', 'Languages' ], true ) ) {
-					continue;
-				}
-
 				if ( empty( $headers[ $value ] ) || false === stripos( $value, 'Plugin' ) ) {
 					continue;
 				}
@@ -235,8 +231,8 @@ class Plugin extends Base {
 	/**
 	 * Add branch switch row to plugins page.
 	 *
-	 * @param $plugin_file
-	 * @param $plugin_data
+	 * @param string    $plugin_file
+	 * @param \stdClass $plugin_data
 	 *
 	 * @return bool
 	 */
@@ -285,8 +281,8 @@ class Plugin extends Base {
 	/**
 	 * Add 'View details' link to plugins page.
 	 *
-	 * @param $links
-	 * @param $file
+	 * @param array  $links
+	 * @param string $file
 	 *
 	 * @return array $links
 	 */
@@ -334,9 +330,9 @@ class Plugin extends Base {
 	/**
 	 * Put changelog in plugins_api, return WP.org data as appropriate
 	 *
-	 * @param $false
-	 * @param $action
-	 * @param $response
+	 * @param bool      $false
+	 * @param string    $action
+	 * @param \stdClass $response
 	 *
 	 * @return mixed
 	 */
@@ -385,7 +381,7 @@ class Plugin extends Base {
 	/**
 	 * Hook into pre_set_site_transient_update_plugins to update from GitHub.
 	 *
-	 * @param $transient
+	 * @param \stdClass $transient
 	 *
 	 * @return mixed
 	 */